@lygztq 2021-04-29
Could you also modify `_segmented_knn_graph_topk` function below? I think both of these two functions need this improvement. Also, @hetong007 I think we can modify docstrings of these functions now (currently, the output of these functions will be copied to CPU regardless of the context of input)