@milesial 2021-04-29
@lygztq I modified the segmented version.

For the docstring, is something like this ok? :

"
If using the 'topk' algorithm, the returned graph is on the same device as input :attr:`x`. Else, the returned graph is on CPU, regardless of the context of the input :attr:`x`.
"