@milesial 2021-04-30
Docstrings updated