@hetong007 2021-05-06
@lygztq Ideally we'd like to have the output knn graph from all algorithms on the same context as the input tensor. Since you are working on a succeeding patch, please remember to fix it as well.

Other than that, LGTM